+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This release adds support to OneView Rest API version 600 for the hardware varia
    - Internal Link Set
    - Logical Enclosure
    - Logical Interconnect
+   - Logical Interconnect Group
    - SAS Interconnect
    - SAS Interconnect Type
    - Server Hardware